refactor: . Move DateAbbreviations to lib/ in src/ and tests/
